+++ NEWS.md
@@ -1,3 +1,5 @@
+# tibblify (development version)
+
 # tibblify 0.4.1
 
 * Fixed type mismatch in C code (#323).